backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.55.3
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pymupdf==1.23.14
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
cryptography==41.0.8